[misc] re-organise folder structure
