+++ src/App.test.js
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom';
 import App from './App';
 
 // mock renderers since webgl methods are not implemented in test framework
-jest.mock('./components/MandelbrotRenderer', () => () => <div />);
-jest.mock('./components/JuliaRenderer', () => () => <div />);
+jest.mock('./components/render/MandelbrotRenderer', () => () => <div />);
+jest.mock('./components/render/JuliaRenderer', () => () => <div />);
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
