+++ src/components/SettingsContext.tsx
@@ -1,8 +1,8 @@
-import React, { ReactNode, useState } from 'react';
+import React, { createContext, ReactNode, useState } from 'react';
 import { SettingsContextType, settingsDefinitionsType } from '../common/settings';
 import defaultSettings, { settingsWidgets } from './SettingsDefinitions';
 
-export const SettingsContext = React.createContext<SettingsContextType>({
+export const SettingsContext = createContext<SettingsContextType>({
   // provide the context with the default values
   settings: defaultSettings,
   // can't set state until it's initialized
@@ -12,7 +12,7 @@ export const SettingsContext = React.createContext<SettingsContextType>({
   settingsWidgets: settingsWidgets(defaultSettings),
 });
 
-// export const useSettings = () => useContext(SettingsContext);
+// export const useSettings = (): SettingsContextType => useContext(SettingsContext);
 
 export const SettingsProvider = ({ children }: { children: ReactNode }): JSX.Element => {
   // https://reactjs.org/docs/context.html#updating-context-from-a-nested-component
